+++ engine/src/parser/components/expression/bexpression.rs
@@ -1,4 +1,4 @@
-use crate::parser::packrat::{PackratParser, ParseSuccess};
+use crate::parser::parser::{PackratParser, ParseSuccess};
 use crate::lexer::token::{CoreToken};
 use crate::errors::{ParseError, SyntaxError, SemanticError, aggregate_errors};
 use crate::constants::common::BOOL;
