+++ engine/src/parser/components/simple_stmt/function_call.rs
@@ -1,4 +1,4 @@
-use crate::parser::packrat::{PackratParser, ParseSuccess};
+use crate::parser::parser::{PackratParser, ParseSuccess};
 use crate::errors::{ParseError, SyntaxError, SemanticError};
 use crate::lexer::token::CoreToken;
