+++ engine/src/parser/parser.rs
@@ -1,5 +1,5 @@
 // Default parser for jarvil uses Packrat approach, first given by Bryan Ford in his master thesis at MIT. It is essentially a 
-// top down recursive descent parsing with lazy memonization in order to avoid exponential parse time and provide reliable 
+// top down recursive descent parsing with lazy memoization in order to avoid exponential parse time and provide reliable 
 // linear time parsing!
 // See https://pdos.csail.mit.edu/~baford/packrat/thesis/ for more information.
